requirements.txt
pymupdf
pillow
torch
transformers
